backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.6.4
email-validator==2.1.1

# ML & Data science — no version pins, pip picks compatible pre-built wheels
scikit-learn
numpy
pandas
fpdf
# PDF generation
reportlab==4.1.0

# HTTP
httpx==0.27.0